Fix some leftover parts
