+++ packages/shared/src/lib/core/activities/utils/outputs/index.ts
@@ -9,7 +9,6 @@ export * from './getNftId'
 export * from './getMetadataFromNftOutput'
 export * from './getRecipientAddressFromOutput'
 export * from './getRecipientFromOutput'
-export * from '../../../wallet/utils/getTimeDifference'
 export * from './isOutputAsync'
 export * from './preprocessGroupedOutputs'
 export * from './preprocessTransaction'
