+++ packages/shared/src/lib/core/wallet/utils/index.ts
@@ -16,7 +16,6 @@ export * from './getMetadataFromFoundryOutput'
 export * from './getOutputParameters'
 export * from './getSenderAddressFromUnlockCondition'
 export * from './getSubjectFromAddress'
-export * from './getTimeDifference'
 export * from './getUnitFromTokenMetadata'
 export * from './hashOutputId'
 export * from './isAccountAddress'
